pyproject.toml
[tool.poetry]
name = "djbugtracker"
version = "0.1.0"
description = ""
authors = ["John Laffey <laffeychem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
django-widget-tweaks = "^1.4.12"
pillow = "^9.4.0"
django-bootstrap5 = "^22.2"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
